test: update test imports and move test files to services/
- Update mock paths: vibe3.utils.* → vibe3.services.*
- Move test files to match new module location:
  - test_git_path_client.py
  - test_handoff_resolution.py
  - test_handoff_security.py
  - test_path_helpers.py
  - test_pr_branch_resolver.py
