+++ tests/vibe3/commands/test_flow_update.py
@@ -22,7 +22,7 @@ runner = CliRunner()
 
 @patch("vibe3.commands.flow_manage.render_flow_created")
 @patch("vibe3.commands.flow_manage.FlowService")
-@patch("vibe3.utils.branch_arg.GitClient")
+@patch("vibe3.services.branch_arg.GitClient")
 def test_flow_update_idempotent(
     git_client_cls,
     flow_service_cls,
@@ -145,7 +145,7 @@ def test_flow_update_blocks_when_branch_has_live_runtime_session(
     registry.get_truly_live_sessions_for_branch.return_value = [{"id": 1}]
 
     with patch(
-        "vibe3.utils.branch_arg.resolve_branch_arg", return_value="task/issue-123"
+        "vibe3.services.branch_arg.resolve_branch_arg", return_value="task/issue-123"
     ):
         result = runner.invoke(flow_app, ["update"])
 
@@ -188,7 +188,7 @@ class TestFlowAddStatusCheck:
     """Tests for flow add status checking."""
 
     @patch("vibe3.commands.flow_manage.FlowService")
-    @patch("vibe3.utils.branch_arg.GitClient")
+    @patch("vibe3.services.branch_arg.GitClient")
     def test_unregistered_branch_creates_flow(self, mock_git_class, mock_service_class):
         """A branch without any flow record should create a new flow."""
         mock_git = MagicMock()
