+++ tests/vibe3/services/test_git_path_client.py
@@ -8,7 +8,7 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from vibe3.utils.git_path_client import (
+from vibe3.services.git_path_client import (
     GitPathProtocol,
     _get_git_client,
     find_worktree_path_for_branch,
