+++ tests/vibe3/services/test_handoff_security.py
@@ -8,7 +8,7 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from vibe3.utils.handoff_resolution import resolve_handoff_target
+from vibe3.services.handoff_resolution import resolve_handoff_target
 
 
 def _make_git_client(git_common: str, worktree_root: str) -> MagicMock:
