+++ tests/vibe3/services/test_path_helpers.py
@@ -1,10 +1,10 @@
 from pathlib import Path
 
-from vibe3.utils.handoff_resolution import (
+from vibe3.services.handoff_resolution import (
     is_shared_handoff_ref,
     to_display_target,
 )
-from vibe3.utils.path_helpers import (
+from vibe3.services.path_helpers import (
     ref_to_handoff_cmd,
     resolve_ref_path,
     sanitize_event_detail_paths,
